next-ml-image-classifier/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
transformers==4.44.0
torch==2.2.2
Pillow==10.4.0
python-multipart==0.0.9
numpy<2
